CNregions.mod Modified CNregions function from the facets package to handle small sample sizes.

CNregions.mod Modified CNregions function from the facets package to handle small sample sizes.

CNregions.mod(
  seg,
  epsilon = 0.005,
  adaptive = FALSE,
  rmCNV = FALSE,
  cnv = NULL,
  frac.overlap = 0.5,
  rmSmallseg = TRUE,
  nProbes = 15
)

Arguments

seg
a segmentation file containing the segmentation information of multiple patients

epsilon
the maximum Euclidean distance between adjacent probes tolerated for denying a nonredundant region. epsilon=0 is equivalent to taking the union of all unique break points across the n samples.

adaptive
Vector of length-m lasso penalty terms.

rmCNV
If TRUE, remove germline CNV.

cnv
A data frame containing germline CNV data.

frac.overlap
A parameter needed to be explain.

rmSmallseg
If TRUE, remove small segment.

nProbes
The segment length threshold below which the segment will be removed if rmSmallseq = TRUE.

Value

reducedM : A matrix ready for plotting.

Examples

library(gnomeR)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
<<<<<<< HEAD
#>     intersect, setdiff, setequal, union

samples <-
as.character(unique(mut$Tumor_Sample_Barcode))[
sample(1:length(unique(mut$Tumor_Sample_Barcode)), 100, replace=FALSE)]
CNregions.mod(seg %>% filter(ID %in% samples))[1:5,1:5]

#>                   chr1.2488138-11167550 chr1.11167550-14192672
=======
#>     intersect, setdiff, setequal, union

samples <-
as.character(unique(mut$Tumor_Sample_Barcode))[
sample(1:length(unique(mut$Tumor_Sample_Barcode)), 100, replace=FALSE)]
CNregions.mod(seg %>% filter(ID %in% samples))

#>                   chr1.2488138-11167550 chr1.11167550-14192672
>>>>>>> a875652248dbaae97b269cfe4efb962be077b5d8
#> P-0000145-T01-IM3               -0.0290                -0.0290
#> P-0000216-T02-IM3               -0.2137                -0.2137
#> P-0000231-T01-IM3               -0.0391                -0.0391
#> P-0000242-T01-IM3               -0.3706                 0.0444
#> P-0000244-T01-IM3               -0.3711                -0.0041
#>                   chr1.14192672-16199471 chr1.16199471-17345415
#> P-0000145-T01-IM3                -0.0290                -0.0290
#> P-0000216-T02-IM3                -0.2137                -0.2137
#> P-0000231-T01-IM3                -0.0391                -0.0391
#> P-0000242-T01-IM3                 0.0444                 0.0444
#> P-0000244-T01-IM3                -0.0041                -0.0041
#>                   chr1.17345415-22587878
#> P-0000145-T01-IM3                -0.0290
#> P-0000216-T02-IM3                -0.2137
#> P-0000231-T01-IM3                -0.0391
#> P-0000242-T01-IM3                 0.0444
#> P-0000244-T01-IM3                -0.0041