custom_pathway Enables creation of a custom pathway binary matrix from a binmat() `object`

custom_pathway Enables creation of a custom pathway binary matrix from a binmat() `object`

custom_pathway(mat, pathway)

Arguments

mat
a binmat() binary matrix

pathway
a dataframe/tibble with first column the name of the genes, and second column their corresponding pathway

Value

pathway_dat : a binary matrix of pathway level alterations

Examples

library(gnomeR)
patients <- as.character(unique(mut$Tumor_Sample_Barcode))[1:200]
bin.mut <- binmat(patients = patients,maf = mut)

#> Warning: All samples were sequenced on the same platform.
#>               The spe.plat argument has been overwritten to FALSE.

pathway <- as.data.frame(cbind(c("PIK3CA","KRAS","TERT","TP53"),c("path1","path1","path2","path3")))
custom_pathway(mat = bin.mut, pathway = pathway)

#>                   path1 path2 path3
#> P-0000062-T01-IM3     0     0     1
#> P-0000065-T01-IM3     1     1     0
#> P-0000113-T01-IM3     0     0     1
#> P-0000140-T01-IM3     0     0     0
#> P-0000145-T01-IM3     1     0     1
#> P-0000157-T01-IM3     1     0     0
#> P-0000185-T01-IM3     1     0     1
#> P-0000088-T01-IM3     0     1     1
#> P-0000193-T01-IM3     0     0     0
#> P-0000230-T01-IM3     1     0     1
#> P-0000231-T01-IM3     0     0     1
#> P-0000242-T01-IM3     0     0     1
#> P-0000244-T01-IM3     0     0     1
#> P-0000270-T01-IM3     0     0     1
#> P-0000281-T01-IM3     0     0     0
#> P-0000300-T01-IM3     1     0     0
#> P-0000329-T01-IM3     0     0     1
#> P-0000336-T01-IM3     0     1     1
#> P-0000340-T01-IM3     0     0     0
#> P-0000363-T01-IM3     0     0     1
#> P-0000370-T01-IM3     0     0     1
#> P-0000372-T01-IM3     1     0     0
#> P-0000407-T01-IM3     0     0     1
#> P-0000426-T01-IM3     1     0     0
#> P-0000427-T01-IM3     0     0     0
#> P-0000447-T01-IM3     0     0     1
#> P-0000497-T01-IM3     0     0     0
#> P-0000510-T01-IM3     0     0     0
#> P-0000530-T01-IM3     1     0     0
#> P-0000538-T01-IM3     0     0     1
#> P-0000546-T01-IM3     0     0     0
#> P-0000569-T01-IM3     0     0     0
#> P-0000573-T01-IM3     0     0     1
#> P-0000625-T01-IM3     1     0     1
#> P-0000678-T01-IM3     1     0     1
#> P-0000684-T01-IM3     0     0     1
#> P-0000708-T01-IM3     0     0     0
#> P-0000726-T01-IM3     0     0     1
#> P-0000745-T01-IM3     1     0     1
#> P-0000761-T01-IM3     0     0     1
#> P-0000770-T01-IM3     0     0     1
#> P-0000777-T01-IM3     0     0     1
#> P-0000789-T01-IM3     0     1     1
#> P-0000791-T01-IM3     1     0     0
#> P-0000788-T01-IM3     1     0     0
#> P-0000795-T01-IM3     0     0     0
#> P-0000805-T01-IM3     0     0     0
#> P-0000809-T01-IM3     0     0     0
#> P-0000838-T01-IM3     0     0     0
#> P-0000842-T01-IM3     0     0     0
#> P-0000847-T01-IM3     0     0     0
#> P-0000855-T01-IM3     0     0     1
#> P-0000861-T01-IM3     0     0     0
#> P-0000820-T02-IM3     0     0     0
#> P-0000878-T01-IM3     0     0     1
#> P-0000879-T01-IM3     0     1     0
#> P-0000904-T01-IM3     0     1     0
#> P-0000912-T01-IM3     0     0     0
#> P-0000914-T01-IM3     0     0     0
#> P-0000929-T01-IM3     0     0     1
#> P-0000954-T01-IM3     0     1     1
#> P-0001001-T01-IM3     0     0     0
#> P-0001004-T01-IM3     0     0     0
#> P-0001027-T01-IM3     0     0     0
#> P-0001046-T01-IM3     0     0     1
#> P-0001038-T01-IM3     0     0     1
#> P-0000955-T01-IM3     0     0     0
#> P-0001096-T01-IM3     0     0     0
#> P-0001103-T01-IM3     0     0     0
#> P-0000963-T01-IM3     0     0     1
#> P-0001107-T01-IM3     1     0     1
#> P-0001158-T01-IM3     0     0     1
#> P-0001177-T01-IM3     0     0     0
#> P-0001181-T01-IM3     0     0     0
#> P-0001197-T01-IM3     0     0     1
#> P-0001208-T01-IM3     1     0     0
#> P-0001218-T01-IM3     1     0     0
#> P-0001239-T01-IM3     0     1     0
#> P-0001255-T01-IM3     1     0     1
#> P-0001286-T01-IM3     1     0     0
#> P-0001296-T01-IM3     0     0     1
#> P-0001303-T01-IM3     0     0     1
#> P-0001293-T01-IM3     0     0     1
#> P-0001320-T01-IM3     0     0     1
#> P-0001337-T01-IM3     0     0     0
#> P-0001350-T01-IM3     1     0     0
#> P-0001345-T01-IM3     0     0     0
#> P-0000964-T02-IM3     0     0     0
#> P-0001385-T01-IM3     1     0     1
#> P-0001397-T01-IM3     0     0     0
#> P-0001388-T01-IM3     0     1     0
#> P-0001427-T01-IM3     1     0     1
#> P-0001460-T01-IM3     0     0     1
#> P-0001485-T01-IM3     0     1     0
#> P-0000216-T02-IM3     1     0     0
#> P-0001493-T01-IM3     0     0     0
#> P-0001507-T01-IM3     0     0     0
#> P-0001512-T01-IM3     0     0     0
#> P-0001538-T01-IM3     0     1     0
#> P-0001540-T01-IM3     0     0     0
#> P-0001555-T01-IM3     0     0     1
#> P-0001551-T01-IM3     0     0     1
#> P-0001564-T01-IM3     0     0     1
#> P-0001583-T01-IM3     1     0     1
#> P-0001596-T01-IM3     0     0     0
#> P-0001585-T01-IM3     0     0     0
#> P-0001603-T01-IM3     0     1     1
#> P-0001611-T01-IM3     0     0     1
#> P-0001616-T01-IM3     0     0     1
#> P-0001358-T02-IM3     1     1     0
#> P-0001669-T01-IM3     0     0     0
#> P-0001687-T01-IM3     0     0     0
#> P-0001680-T02-IM3     0     1     0
#> P-0001701-T01-IM3     0     0     1
#> P-0001711-T01-IM3     0     0     0
#> P-0001724-T01-IM3     0     0     0
#> P-0001730-T01-IM3     0     0     0
#> P-0001741-T01-IM3     1     0     1
#> P-0001752-T01-IM3     0     0     1
#> P-0001781-T01-IM3     0     0     0
#> P-0001787-T01-IM3     0     0     1
#> P-0001822-T01-IM3     0     0     0
#> P-0001823-T01-IM3     0     0     1
#> P-0001830-T01-IM3     1     0     1
#> P-0001861-T01-IM3     0     0     1
#> P-0001884-T01-IM3     0     0     0
#> P-0001919-T01-IM3     0     0     0
#> P-0001930-T01-IM3     0     0     0
#> P-0001991-T01-IM3     0     1     0
#> P-0002021-T01-IM3     0     0     1
#> P-0001684-T02-IM3     0     0     0
#> P-0002060-T01-IM3     0     0     1
#> P-0002063-T01-IM3     0     1     0
#> P-0002069-T01-IM3     1     0     0
#> P-0000715-T03-IM3     0     1     0
#> P-0002091-T01-IM3     0     0     1
#> P-0002096-T01-IM3     0     0     0
#> P-0001135-T02-IM3     0     1     1
#> P-0002127-T01-IM3     0     0     0
#> P-0002135-T01-IM3     0     0     0
#> P-0002144-T01-IM3     0     0     0
#> P-0002141-T01-IM3     0     1     0
#> P-0002160-T01-IM3     0     0     1
#> P-0002164-T01-IM3     0     0     1
#> P-0002207-T01-IM3     0     0     1
#> P-0000377-T03-IM3     0     0     1
#> P-0002240-T01-IM3     0     0     1
#> P-0002257-T01-IM3     0     0     1
#> P-0002291-T01-IM3     1     0     1
#> P-0002292-T01-IM3     0     1     1
#> P-0002296-T01-IM3     0     0     0
#> P-0002313-T01-IM3     0     0     1
#> P-0002327-T01-IM3     1     0     1
#> P-0002343-T01-IM3     0     0     1
#> P-0002344-T01-IM3     0     1     1
#> P-0002273-T02-IM3     0     0     0
#> P-0002367-T01-IM3     0     1     1
#> P-0002374-T01-IM3     0     0     1
#> P-0002396-T01-IM3     1     1     0
#> P-0002405-T01-IM3     0     1     1
#> P-0002407-T01-IM3     0     0     1
#> P-0002421-T01-IM3     1     0     1
#> P-0002434-T01-IM3     1     0     0
#> P-0002442-T01-IM3     0     0     1
#> P-0002456-T01-IM3     1     0     0
#> P-0002463-T01-IM3     1     0     0
#> P-0002500-T01-IM3     0     0     0
#> P-0002510-T01-IM3     0     1     0
#> P-0002536-T01-IM3     0     0     1
#> P-0002533-T01-IM3     0     0     0
#> P-0002580-T01-IM3     0     0     0
#> P-0002591-T01-IM3     1     0     1
#> P-0002605-T01-IM3     0     0     1
#> P-0002614-T01-IM3     1     0     0
#> P-0002608-T01-IM3     0     0     0
#> P-0002616-T01-IM3     1     0     0
#> P-0002624-T01-IM3     0     0     0
#> P-0002629-T01-IM3     0     0     1
#> P-0002651-T01-IM3     0     0     1
#> P-0002656-T01-IM3     0     0     0
#> P-0002658-T01-IM3     0     0     1
#> P-0002673-T01-IM3     1     0     1
#> P-0001942-T03-IM3     0     0     1
#> P-0002640-T01-IM3     0     0     0
#> P-0002706-T01-IM3     0     1     0
#> P-0002729-T01-IM3     0     0     1
#> P-0002770-T01-IM3     0     0     1
#> P-0002777-T01-IM3     0     1     1
#> P-0002783-T01-IM3     0     0     1
#> P-0002794-T01-IM3     0     0     0
#> P-0002828-T01-IM3     0     0     0
#> P-0002830-T01-IM3     1     0     0
#> P-0002844-T01-IM3     0     0     0
#> P-0002881-T01-IM3     0     0     0
#> P-0002915-T01-IM3     1     0     0
#> P-0002921-T01-IM3     1     0     0
#> P-0002940-T01-IM3     1     0     0
#> P-0002955-T01-IM3     1     0     0
#> P-0002750-T02-IM3     0     0     0
#> P-0002960-T01-IM3     0     0     1