gen.summary

Creates a summary table of the distribution of the genetic features by a specific outcome/covariate of interest. The difference is tested using Fisher's exact test and further adjusted for multiple comparisons. Note that continuous genetic factors are dichotomized at their median.

gen.summary(gen.dat, outcome, filter = 0, cont = F, rank = T, na.filt = 0)

Arguments

gen.dat

A matrix or dataframe, with patients as rows and features as columns.

outcome

A leveled vector of length equal to the number of rows in gen.dat.

filter

a numeric value between 0 and 1 (1 not included) that is the lower bound for the proportion of patients having a genetic event (only for binary features). All features with an event rate lower than that value will be removed. Default is 0 (all features included).

cont

Should the outcome be treated as a continuous value. Default is FALSE treated as categorical.

rank

Should the table returned be ordered by Pvalue. Boolean, default is T

na.filt

A numeric value between 0 and 1 (1 not included) that is the upper bound for the proportion of missing values in the features of the inputted gen.dat matrix. Variables that exceed this proportion of missing values will be removed.

Value

fits : a table of odds ratio and pvalues. forest.plot : A forest plot of the top 10 hits.

Examples

library(gnomeR)
patients <- as.character(unique(mut$Tumor_Sample_Barcode))
## binary outcome ##
outcome <- as.character(clin.sample$Sample.Type[match(patients,clin.sample$Sample.Identifier)])
gen.dat <- binmat(patients = patients,maf = mut)
#> Warning: MUTATION DATA: To ensure gene with multiple names/aliases are correctly grouped together, the
#>       following genes in your maf dataframe have been recoded. You can supress this with recode.aliases = FALSE 
#>  
#> MLL3 recoded to 58508 
#> MUTATION DATA: To ensure gene with multiple names/aliases are correctly grouped together, the
#>       following genes in your maf dataframe have been recoded. You can supress this with recode.aliases = FALSE 
#>  
#> MLL2 recoded to 8085 
#> MUTATION DATA: To ensure gene with multiple names/aliases are correctly grouped together, the
#>       following genes in your maf dataframe have been recoded. You can supress this with recode.aliases = FALSE 
#>  
#> MRE11A recoded to 4361 
#> MUTATION DATA: To ensure gene with multiple names/aliases are correctly grouped together, the
#>       following genes in your maf dataframe have been recoded. You can supress this with recode.aliases = FALSE 
#>  
#> MLL recoded to 4297 
#> MUTATION DATA: To ensure gene with multiple names/aliases are correctly grouped together, the
#>       following genes in your maf dataframe have been recoded. You can supress this with recode.aliases = FALSE 
#>  
#> PARK2 recoded to 5071 
#> MUTATION DATA: To ensure gene with multiple names/aliases are correctly grouped together, the
#>       following genes in your maf dataframe have been recoded. You can supress this with recode.aliases = FALSE 
#>  
#> RFWD2 recoded to 64326 
#> MUTATION DATA: To ensure gene with multiple names/aliases are correctly grouped together, the
#>       following genes in your maf dataframe have been recoded. You can supress this with recode.aliases = FALSE 
#>  
#> PAK7 recoded to 57144 
#> MUTATION DATA: To ensure gene with multiple names/aliases are correctly grouped together, the
#>       following genes in your maf dataframe have been recoded. You can supress this with recode.aliases = FALSE 
#>  
#> FAM46C recoded to 54855 
#> MUTATION DATA: To ensure gene with multiple names/aliases are correctly grouped together, the
#>       following genes in your maf dataframe have been recoded. You can supress this with recode.aliases = FALSE 
#>  
#> FAM175A recoded to 84142 
#> MUTATION DATA: To ensure gene with multiple names/aliases are correctly grouped together, the
#>       following genes in your maf dataframe have been recoded. You can supress this with recode.aliases = FALSE 
#>  
#> MYCL1 recoded to 4610 
#> MUTATION DATA: To ensure gene with multiple names/aliases are correctly grouped together, the
#>       following genes in your maf dataframe have been recoded. You can supress this with recode.aliases = FALSE 
#>  
#> TCEB1 recoded to 6921 
#> Warning: When panels argument is NULL only IMPACT samples can be annotated for missingness.
#>           If you are using GENIE data, we recommend that you provide this argument.
## Continuous outcome ##
set.seed(1)
outcome <-  rnorm(n = nrow(gen.dat))
tab.out <- gen.summary(gen.dat = gen.dat,
                   outcome = outcome,
                   filter = 0.05,
                   cont = TRUE,rank = TRUE)
tab.out$fits
#>          Feature EventFrequency Estimate    SD  Lower Upper   Pvalue      FDR
#> PIK3CA.x PIK3CA           0.107   -0.235 0.149 -0.526 0.056 1.14e-01 9.46e-01
#> RB1.x       RB1           0.050   -0.291 0.210 -0.704 0.121 1.67e-01 9.46e-01
#> TERT.x     TERT           0.149    0.165 0.129 -0.088 0.418 2.03e-01 9.46e-01
#> BRAF.x     BRAF           0.050    0.226 0.211 -0.187 0.639 2.83e-01 9.77e-01
#> TP53.x     TP53           0.429    0.073 0.093 -0.109 0.256 4.31e-01 9.77e-01
#> ATRX.x     ATRX           0.057    0.125 0.199 -0.265 0.515 5.30e-01 9.77e-01
#> NOTCH1.x NOTCH1           0.055   -0.085 0.203 -0.482 0.312 6.75e-01 9.77e-01
#> KRAS.x     KRAS           0.127    0.056 0.138 -0.216 0.327 6.87e-01 9.77e-01
#> CDKN2A.x CDKN2A           0.053    0.061 0.207 -0.344 0.466 7.68e-01 9.77e-01
#> APC.x       APC           0.079   -0.039 0.171 -0.375 0.296 8.18e-01 9.77e-01
#> ARID1A.x ARID1A           0.079   -0.034 0.171 -0.370 0.301 8.40e-01 9.77e-01
#> EGFR.x     EGFR           0.085   -0.019 0.165 -0.343 0.304 9.06e-01 9.77e-01
#> ATM.x       ATM           0.055    0.018 0.203 -0.380 0.415 9.30e-01 9.77e-01
#> FAT1.x     FAT1           0.061   -0.005 0.192 -0.382 0.371 9.77e-01 9.77e-01
tab.out$vPlot