gen.tab

Creates a summary table of the distribution of the genetic features by a specific outcome/covariate of interest. The difference is tested using Fisher's exact test and further adjusted for multiple comparisons. Note that continuous genetic factors are dichotomized at their median.

gen.tab(gen.dat, outcome, filter = 0, paired = F, cont = F, rank = T)

Arguments

gen.dat
A matrix or dataframe, with patients as rows and features as columns.

outcome
A leveled vector of length equal to the number of rows in gen.dat.

filter
a numeric value between 0 and 1 (1 not included) that is the lower bound for the proportion of patients having a genetic event (only for binary features). All features with an event rate lower than that value will be removed. Default is 0 (all features included).

paired
Boolean if the data are paired. Default is FALSE.

cont
Should the outcome be treated as a continuous value. Default is FALSE treated as categorical.

rank
Should the table returned be ordered by Pvalue. Boolean, default is T

Value

fits : a table of odds ratio and pvalues.

forest.plot : A forest plot of the top 10 hits.

Examples

library(gnomeR)
patients <- unique(clin$DMPID)[1:500]

#> Error in unique(clin$DMPID): object 'clin' not found

mut.only <- create.bin.matrix(patients = patients,maf = mut)

#> Error in create.bin.matrix(patients = patients, maf = mut): could not find function "create.bin.matrix"

gen.dat <- mut.only$mut

#> Error in eval(expr, envir, enclos): object 'mut.only' not found

outcome = as.factor(as.character(clin$Sex[match(patients,clin$DMPID)]))

#> Error in is.factor(x): object 'clin' not found

test <- gen.tab(gen.dat,outcome)

#> Error in apply(gen.dat, 2, function(x) {    length(unique(x[!is.na(x)])) == 1} || all(is.na(x))): object 'gen.dat' not found

head(test$fits)

#> Error in head(test$fits): object 'test' not found

test$forest.plot

#> Error in eval(expr, envir, enclos): object 'test' not found