gen_summary

Creates a summary table of the distribution of the genetic features by a specific outcome/covariate of interest. The difference is tested using Fisher's exact test and further adjusted for multiple comparisons. Note that continuous genetic factors are dichotomized at their median.

gen_summary(gen_dat, outcome, filter = 0, cont = F, rank = T, na_filter = 0)

Arguments

gen_dat

A matrix or dataframe, with samples as rows and features as columns.

outcome

A leveled vector of length equal to the number of rows in gen_dat.

filter

a numeric value between 0 and 1 (1 not included) that is the lower bound for the proportion of samples having a genetic event (only for binary features). All features with an event rate lower than that value will be removed. Default is 0 (all features included).

cont

Should the outcome be treated as a continuous value. Default is FALSE treated as categorical.

rank

Should the table returned be ordered by Pvalue. Boolean, default is T

na_filter

A numeric value between 0 and 1 (1 not included) that is the upper bound for the proportion of missing values in the features of the inputted gen_dat matrix. Variables that exceed this proportion of missing values will be removed.

Value

fits : a table of odds ratio and pvalues. forest.plot : A forest plot of the top 10 hits.

Examples

library(gnomeR)
samples <- as.character(unique(mut$Tumor_Sample_Barcode))
## binary outcome ##
outcome <- as.character(clin.sample$Sample.Type[match(samples,clin.sample$Sample.Identifier)])
gen_dat <- binary_matrix(samples = samples,mutation = mut)
## Continuous outcome ##
set.seed(1)
outcome <-  rnorm(n = nrow(gen_dat))
tab.out <- gen_summary(gen_dat = gen_dat,
                   outcome = outcome,
                   filter = 0.05,
                   cont = TRUE, rank = TRUE)
tab.out$fits
#>          Feature EventFrequency Estimate    SD  Lower Upper   Pvalue      FDR
#> TERT.x     TERT           0.149    0.258 0.129  0.005 0.510 4.63e-02 6.78e-01
#> BRAF.x     BRAF           0.050    0.291 0.210 -0.121 0.703 1.67e-01 6.78e-01
#> PIK3CA.x PIK3CA           0.107   -0.197 0.149 -0.488 0.095 1.87e-01 6.78e-01
#> RB1.x       RB1           0.050   -0.270 0.210 -0.682 0.143 2.01e-01 6.78e-01
#> ATM.x       ATM           0.055   -0.237 0.202 -0.634 0.160 2.42e-01 6.78e-01
#> KRAS.x     KRAS           0.127   -0.132 0.138 -0.403 0.140 3.42e-01 6.78e-01
#> EGFR.x     EGFR           0.085    0.131 0.165 -0.192 0.454 4.28e-01 6.78e-01
#> MLL2.x     MLL2           0.070    0.133 0.181 -0.221 0.487 4.62e-01 6.78e-01
#> CDKN2A.x CDKN2A           0.053    0.150 0.207 -0.255 0.555 4.67e-01 6.78e-01
#> MLL3.x     MLL3           0.072   -0.119 0.178 -0.468 0.230 5.03e-01 6.78e-01
#> APC.x       APC           0.079   -0.112 0.171 -0.447 0.224 5.14e-01 6.78e-01
#> TP53.x     TP53           0.429   -0.059 0.093 -0.242 0.123 5.26e-01 6.78e-01
#> ARID1A.x ARID1A           0.079   -0.102 0.171 -0.437 0.233 5.51e-01 6.78e-01
#> ATRX.x     ATRX           0.057   -0.094 0.199 -0.484 0.296 6.37e-01 7.28e-01
#> FAT1.x     FAT1           0.061    0.058 0.192 -0.319 0.434 7.64e-01 8.15e-01
#> NOTCH1.x NOTCH1           0.055    0.036 0.203 -0.361 0.433 8.59e-01 8.59e-01
tab.out$vPlot