Get all genetic data for a study or list of IDs

Get all genetic data for a study or list of IDs

get_genetics(
  sample_ids = NULL,
  sample_list_id = NULL,
  genes = "all_impact",
  database = "msk_impact",
  mutations = TRUE,
  fusions = TRUE,
  cna = TRUE
)

Arguments

sample_ids
A character vector of sample ids

sample_list_id
A character vector naming a pre-specified list of samples (e.g. `"mskimpact_Colorectal_Cancer"`)

genes
A list of genes to query. default is all impact genes.

database
A character string of the database to be used. Options are "msk_impact" or "tcga", default is "msk_impact".

mutations
Boolean specifying if mutation data should be fected. Default is TRUE.

fusions
Boolean specifying if fusion data should be fected. Default is TRUE.

cna
Boolean specifying if cna data should be fected. Default is TRUE.

Value

A dataframe of mutations for each sample ID

Examples

if (FALSE) {
get_genetics(sample_ids = c("P-0000004-T01-IM3", "P-0000012-T02-IM3"),
genes = 207)
}