List of suggested color palettes for when you need a large palette

Sometimes you just need a huge palette of fairly distinguishable colors. This is a named vector of Ronglai-approved colors good for things like TCGA PanCan (33 cancer types) or clustering solutions with high K. Run gnomer_colors to see the hex codes for the study colors.

Usage

gnomer_colors

Format

An object of class character of length 66.