Creates the OncoPrint corresponding to the inputted genetic data

Creates the OncoPrint corresponding to the inputted genetic data

plot_oncoprint(
  gen_dat,
  clin_dat = NULL,
  ordered_samples = NULL,
  rank_genes = TRUE,
  background_color = "#CCCCCC"
)

Arguments

gen_dat

A binary matrix or dataframe, with samples as rows and features as columns. Note that the names of the columns must end in ".Del" or ".Amp" to recognize copy number alterations. (see create.bin.matrix for more details on this format).

clin_dat

An optional clinical file, including only the features the user wishes to add to the plot. Default is NULL.

ordered_samples

An optional vector of length equal to the number of samples under consideration. Indicates the new order (from left to right) to be plotted.

rank_genes

Boolean argument specifying if the genes should be ranked by event frequency. Default is TRUE.

background_color

Color name to be used to fill the background of the OncoPrint.

Value

an oncoprint object

Examples

library(gnomeR)