Simplify binary matrix to one column per gene that counts any alteration type as 1

This will reduce the number of columns in your binary matrix, and the resulting data frame will have only 1 col per gene, as opposed to separate columns for mutation/cna/fusion.

summarize_by_gene(gene_binary)

Arguments

gene_binary

a 0/1 matrix of gene alterations

Value

a binary matrix with a row for each sample and one column per gene

Examples

library(gnomeR)
samples <- unique(mutations$sample_id)[1:10]
#> Warning: Unknown or uninitialised column: `sample_id`.
gene_binary <- create_gene_binary(
  samples = samples, mutation = mutations, cna = cna,
  mut_type = "somatic_only",
  include_silent = FALSE,
  specify_panel = "IMPACT341"
) %>%
  summarize_by_gene()
#> ℹ `samples` argument is `NULL`. We will infer your cohort inclusion and resulting data frame will include all samples with at least one alteration in mutation, fusion or cna data frames