CNregions.mod Modified CNregions function from the facets package to handle small sample sizes.

CNregions.mod Modified CNregions function from the facets package to handle small sample sizes.

CNregions.mod(
  seg,
  epsilon = 0.005,
  adaptive = FALSE,
  rmCNV = FALSE,
  cnv = NULL,
  frac.overlap = 0.5,
  rmSmallseg = TRUE,
  nProbes = 15
)

Arguments

seg

a segmentation file containing the segmentation information of multiple samples

epsilon

the maximum Euclidean distance between adjacent probes tolerated for denying a nonredundant region. epsilon=0 is equivalent to taking the union of all unique break points across the n samples.

adaptive

Vector of length-m lasso penalty terms.

rmCNV

If TRUE, remove germline CNV.

cnv

A data frame containing germline CNV data.

frac.overlap

A parameter needed to be explain.

rmSmallseg

If TRUE, remove small segment.

nProbes

The segment length threshold below which the segment will be removed if rmSmallseq = TRUE.

Value

reducedM : A matrix ready for plotting.

Examples

library(gnomeR)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
samples <-
as.character(unique(mut$Tumor_Sample_Barcode))[
sample(1:length(unique(mut$Tumor_Sample_Barcode)), 100, replace=FALSE)]
CNregions.mod(seg %>% filter(ID %in% samples))[1:5,1:5]
#>                   chr1.2488138-9770584 chr1.9770584-9784407
#> P-0000080-T01-IM3               0.1680               0.1680
#> P-0000096-T01-IM3               0.0798               0.0798
#> P-0000157-T01-IM3               0.0431               0.0431
#> P-0000281-T01-IM3              -0.0626              -0.0626
#> P-0000372-T01-IM3               0.3203               0.3203
#>                   chr1.9784407-11167550 chr1.11167550-16235896
#> P-0000080-T01-IM3                0.1680                 0.1680
#> P-0000096-T01-IM3                0.0798                 0.0798
#> P-0000157-T01-IM3                0.0431                 0.0431
#> P-0000281-T01-IM3               -0.0626                -0.0626
#> P-0000372-T01-IM3                0.3203                 0.3203
#>                   chr1.16235896-22587878
#> P-0000080-T01-IM3                 0.1680
#> P-0000096-T01-IM3                 0.0798
#> P-0000157-T01-IM3                 0.0431
#> P-0000281-T01-IM3                -0.0626
#> P-0000372-T01-IM3                 0.3203