CNregions.mod Modified CNregions function from the facets package to handle small sample sizes.

CNregions.mod Modified CNregions function from the facets package to handle small sample sizes.

CNregions.mod(
  seg,
  epsilon = 0.005,
  adaptive = FALSE,
  rmCNV = FALSE,
  cnv = NULL,
  frac.overlap = 0.5,
  rmSmallseg = TRUE,
  nProbes = 15
)

Arguments

seg

a segmentation file containing the segmentation information of multiple samples

epsilon

the maximum Euclidean distance between adjacent probes tolerated for denying a nonredundant region. epsilon=0 is equivalent to taking the union of all unique break points across the n samples.

adaptive

Vector of length-m lasso penalty terms.

rmCNV

If TRUE, remove germline CNV.

cnv

A data frame containing germline CNV data.

frac.overlap

A parameter needed to be explain.

rmSmallseg

If TRUE, remove small segment.

nProbes

The segment length threshold below which the segment will be removed if rmSmallseq = TRUE.

Value

reducedM : A matrix ready for plotting.

Examples

library(gnomeR)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
samples <-
as.character(unique(mutations$sampleId))[
sample(1:length(unique(mutations$sampleId)), 100, replace=FALSE)]
CNregions.mod(seg %>% filter(ID %in% samples))[1:5,1:5]
#>                   chr1.2488138-9784407 chr1.9784407-11168291
#> P-0000842-T01-IM3               0.0748                0.0748
#> P-0000847-T01-IM3              -0.0039               -0.0039
#> P-0000964-T02-IM3               0.0023                0.0023
#> P-0001337-T01-IM3              -0.2677               -0.2677
#> P-0001493-T01-IM3              -0.0079               -0.0079
#>                   chr1.11168291-11175489 chr1.11175489-14192672
#> P-0000842-T01-IM3                 0.0748                 0.0748
#> P-0000847-T01-IM3                -0.0039                -0.0039
#> P-0000964-T02-IM3                 0.0023                 0.0023
#> P-0001337-T01-IM3                -0.2677                 0.0441
#> P-0001493-T01-IM3                -0.0079                -0.0079
#>                   chr1.14192672-40361361
#> P-0000842-T01-IM3                 0.0748
#> P-0000847-T01-IM3                -0.0039
#> P-0000964-T02-IM3                 0.0023
#> P-0001337-T01-IM3                 0.0441
#> P-0001493-T01-IM3                -0.0079