Pathway Alterations

Checks if certain oncogenic signaling pathways are altered. Pathways were curated from this paper. Please check for gene aliases before using.

add_pathways(
  gene_binary,
  pathways = c(names(gnomeR::pathways)),
  custom_pathways = NULL,
  bind_pathways = TRUE,
  count_pathways_by = c("alteration", "gene")
)

Arguments

gene_binary

a binary matrix from gene_binary()

pathways

a vector of pathway names to annotate. The options are names(gnomeR::pathways) ("RTK/RAS", "Nrf2", "PI3K", "TGFB", "p53", "Wnt", "Myc", "Cell cycle", "Hippo", "Notch"). You can pass one pathway name, multiple pathway names, or NULL. By default, all pathways defined in gnomeR::pathways will be included. Included default pathways are alteration-specific meaning a specific type of alteration (mut/cna/fusion) is required to mark a 1 for that pathway. If you want gene-level pathways, use summarize_by_gene() on your binary matrix, then add pathways you want as custom_pathways

custom_pathways

a vector of alterations to annotate as a pathway, or a list of custom pathways (see gnomeR::pathways as example)

bind_pathways

a logical indicating whether pathway columns should be joined to main gene_binary or returned separately as a list item. Default is TRUE and function will bind_cols() and return a data.frame. If FALSE a list will be returned.

count_pathways_by

Must be one of the following: "alteration" (default), or "gene". This determines whether any type of gene alteration should be counted towards a pathway ("gene") or only specific types of alterations should be counted towards a pathway ("alteration") By default, the function assumes alteration-specific pathway annotation and all default pathways are annotated this way. If a custom pathway is passed with no suffix (e.g. custom_pathway = 'TP53') it will assume it is a mutation.

Value

a data frame: each sample is a row, columns are pathways, with values of 0/1 depending on pathway alteration status.

Details

Sanchez-Vega, F., Mina, M., Armenia, J., Chatila, W. K., Luna, A., La, K. C., Dimitriadoy, S., Liu, D. L., Kantheti, H. S., Saghafinia, S., Chakravarty, D., Daian, F., Gao, Q., Bailey, M. H., Liang, W. W., Foltz, S. M., Shmulevich, I., Ding, L., Heins, Z., Ochoa, A., … Schultz, N. (2018). Oncogenic Signaling Pathways in The Cancer Genome Atlas. Cell, 173(2), 321–337.e10. https://doi.org/10.1016/j.cell.2018.03.035

Examples

gene_binary <- create_gene_binary(mutation = mut, cna = cna, fusion = fusion)
#> ℹ `samples` argument is `NULL`. We will infer your cohort inclusion and resulting data frame will include all samples with at least one alteration in mutation, fusion or cna data frames
pathway_df <- add_pathways(gene_binary, pathways = "Notch")