tbl_genomic

This function will select genes based on user inputs or on frequency counts and then will pass the data.frame to gtsummary::tbl_summary(). You can specify a by variable and other parameters that are accepted by gtsummary::tbl_summary(). Note the by variable must be merged on to onto the data before using the by parameter in the function.

tbl_genomic(
  gene_binary,
  freq_cutoff = 0,
  freq_cutoff_by_gene = TRUE,
  gene_subset = NULL,
  by = NULL,
  ...
)

Arguments

gene_binary

data.frame of genetic samples

freq_cutoff

A number 0 to 1 representing the minimum percent frequency you are using to select gene's for analysis. Frequencies can be calculated at gene level, or alteration level (see freq_cutoff_by_gene).

freq_cutoff_by_gene

Logical indicating whether gene selection based on frequency % should be calculated at the gene level, or the alteration level. Default is TRUE, indicating all alterations (e.g. TP53/TP53.Del/TP53.Del/TP53.fus) will be aggregated at gene level to determine cutoff frequencies and all alteration types for top genes will be returned.

gene_subset

Specific genes you want to summarize. This takes precedent over the freq_cutoff parameter and will return all alterations associated with that gene (including mutations, CNA, fusions).

by

A variable to be passed to gtsummary::tbl_summary()'s by parameter

...

Additional parameters that can be passed to gtsummary::tbl_summary(). To access the additional parameters you need to load gtsummary.

Value

A tbl_summary() object

Examples

library(dplyr)

samples <- unique(mutations$sampleId)[1:10]

gene_binary <- create_gene_binary(
  samples = samples,
  mutation = gnomeR::mutations,
  cna = gnomeR::cna,
  mut_type = "somatic_only", snp_only = FALSE,
  specify_panel = "no"
)
tb1 <- tbl_genomic(gene_binary = gene_binary, freq_cutoff = .05)
tb2 <- tbl_genomic(gene_binary = gene_binary, gene_subset = c("BRCA1", "FGFR2"))

gene_binary <- gene_binary %>%
  mutate(sex = sample(
    x = c("M", "F"),
    size = nrow(gene_binary), replace = TRUE
  ))
library(gtsummary) # Need to load gtsummary to access additional arguments in `...`
#> #StandWithUkraine

t1 <- tbl_genomic(
  gene_binary = gene_binary,
  by = sex,
  freq_cutoff = .2,
  freq_cutoff_by_gene = FALSE,
  statistic = list(all_categorical() ~"{n}")
)